feat(certificate): reorganized icons and create utils file
